unit tests

+++ ModuleTesting/ModuleTesting/GeographicCalculator.cs
@@ -12,8 +12,9 @@ namespace ModuleTesting
         public const double sb = b * b;
         public const double seccentricity = (sa - sb) / sa;
         public const double seccentricity2 = (sa - sb) / sb;
-        /// підрахунок довжини в км. до одного градусу на меридіані
         
+        
+        /// підрахунок довжини в км. до одного градусу на меридіані
         public static double MeridianLength(double lattitude, double d = 1.0)
         {
             if (d > 1) 
@@ -26,13 +27,9 @@ namespace ModuleTesting
             else if (lattitude < 0.0) 
                 lattitude = 0.0;
 
-            double lttM = lattitude + d / 2.0;
-            // взяли середину між нижньою і верхньою частиною дуги
-            // double eccentricity = Math.Sqrt ( seccentricity );
+            double lttM = lattitude + ((double)d) / 2.0;
             double sinB = sin(lttM);
             double foo = Math.Sqrt(1 - seccentricity * sinB * sinB);
-            //double N = a / ( Math.Sqrt (foo));
-            // радиус кривизни першого вертикалу
             double M = a * (1 - seccentricity)
                 / (foo * foo * foo);
             double rc = M * (Math.PI / 180.0); // length of 1 degree
